Throw on not supported upsert & getOrCreate methods
These were already not running as expected and
even after trying to re-implement them I didn't
manage to reach a non-hacky result in a sane
amount of time. On top of that, given that this
library is primarily targeting testing endpoints,
those two methods should be avoided since they
provide an abstraction on top of two other calls.

Change-type: patch
Signed-off-by: Thodoris Greasidis <thodoris@balena.io>